Add IntelliJ module files
For developers using the IntelliJ IDEA IDE,
this better ensures that JDK 1.8/8 is used
for development. This is required for the
Gradle version in use.

It also improves the developer experience
by including run configurations for quickly
testing the client and server versions of
Minecraft with the mod.

Information about the Minecraft plugin for
IntelliJ is also included to improve the
Minecraft plugin setup for developers.